+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- [**BC**] Add `$initiator` argument to `ResponseDecoderInterface::supports` method
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation