+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 - Add `OnErrorCallback::ignoreError` named constructor
+- [**BC**] Add `$initiator` argument to `ResponseDecoderInterface::supports` method
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation