+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Copy `$response` in `ProfilerItem` not to store a reference to the original object
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation