+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 - Copy `$response` in `ProfilerItem` not to store a reference to the original object
+- Add `OnErrorCallback::ignoreError` named constructor
 
 ## 1.0.0 - 2021-05-12
 - Initial implementation